requirements.txt
streamlit
nltk
scikit-learn
pdfplumber
python-docx
